Initial import of ortin, an IS-NITRO-EMULATOR management tool.
Currently, it can load a ROM (with encrypted Secure Area and S-Boxes),
but it doesn't load the debug ROM, so it doesn't actually start. It
ends up showing the DS menu, but you can't start a game directly in
the menu, so it's not very useful.

The banner *does* show up, though.