[libortin] Use the original IS-NITRO-DEBUGGER debugger ROM.
The debug pointers are identical to the NitroDriver ROM. The main issue
was we needed to resume the CPUs from break after initializing everything.

waitForDebuggerROM(): New function to wait for the debugger ROM to
initialize.

sendCpuCMD174(): Send cmd174 to a CPU. Not sure what this does, but
LibISNitroEmulator sends it after the debugger ROM is initialized.